Implement Navigator Interface Refactoring - Phase 1
- Created PaneNavigator extension interface for pane operations
- Created TransitionController internal interface for animations
- Created ResultCapable internal interface for result passing
- Created InternalQuoVadisApi annotation for opt-in internal APIs
- Updated Navigator to slim down to 14 core public members (from 26)
- Updated TreeNavigator and FakeNavigator to implement all new interfaces
- Updated NavigatorResultExtensions to use ResultCapable interface
- All implementations backward compatible via extension interfaces
- Main Navigator interface reduced by 46%
