+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/navigation/tree/config/PopBehavior.kt
@@ -1,0 +1,18 @@
+package com.jermey.quo.vadis.core.navigation.tree.config
+
+/**
+ * Configures behavior when a StackNode becomes empty after pop.
+ */
+enum class PopBehavior {
+    /**
+     * Remove the empty stack from parent (cascading pop).
+     * This continues popping until a non-empty container is found.
+     */
+    CASCADE,
+
+    /**
+     * Preserve the empty stack in place.
+     * The stack remains but with no children.
+     */
+    PRESERVE_EMPTY
+}
