+++ quo-vadis-core/src/commonMain/kotlin/com/jermey/quo/vadis/core/navigation/tree/result/BackResult.kt
@@ -1,0 +1,17 @@
+package com.jermey.quo.vadis.core.navigation.tree.result
+
+import com.jermey.quo.vadis.core.navigation.NavNode
+
+/**
+ * Result of a tree-aware back operation.
+ */
+sealed class BackResult {
+    /** Back was handled, new tree state returned */
+    data class Handled(val newState: NavNode) : BackResult()
+
+    /** Back should be delegated to system (e.g., close app) */
+    data object DelegateToSystem : BackResult()
+
+    /** Back could not be handled (internal error) */
+    data object CannotHandle : BackResult()
+}
