--- quo-vadis-core/src/iosMain/kotlin/com/jermey/quo/vadis/core/compose/PlatformDefaults.ios.kt
@@ -1,2 +1,0 @@
-package com.jermey.quo.vadis.core.compose
-
