--- quo-vadis-core/src/jsMain/kotlin/com/jermey/quo/vadis/core/compose/PlatformDefaults.js.kt
@@ -1,2 +1,0 @@
-package com.jermey.quo.vadis.core.compose
-
