--- quo-vadis-core/src/wasmJsMain/kotlin/com/jermey/quo/vadis/core/compose/PlatformDefaults.wasmJs.kt
@@ -1,2 +1,0 @@
-package com.jermey.quo.vadis.core.compose
-
